Moved all files to their own directories
