--- Caretaker.cpp
@@ -1,9 +1,0 @@
-#include "Caretaker.h"
-
-Caretaker::storeMemento(Memento* memento) {
-    this->canvasses.push_front(memento);
-}
-
-Caretaker::retrieveMemento() {
-    return this->canvasses.pop_front();
-}
