+++ ExportCanvas/ExportCanvas.h
@@ -1,7 +1,7 @@
 #ifndef ExportCanvas_H
 #define ExportCanvas_H
 
-#include "Canvas.h"
+#include "Canvas/Canvas.h"
 
 class ExportCanvas : public Canvas
 {
