+++ PNGExporter/PNGExporter.h
@@ -1,7 +1,7 @@
 #ifndef PNGExporter_H
 #define PNGExporter_H
 
-#include "ExportCanvas.h"
+#include "ExportCanvas/ExportCanvas.h"
 
 class PNGExporter : public ExportCanvas
 {
