+++ SquareFactory/SquareFactory.cpp
