+++ SquareFactory/SquareFactory.h
@@ -1,8 +1,8 @@
 #ifndef SquareFactory_H
 #define SquareFactory_H
 
-#include "ShapeFactory.h"
-#include "Square.h"
+#include "ShapeFactory/ShapeFactory.h"
+#include "Square/Square.h"
 
 class SquareFactory : public ShapeFactory
 {
